refactor(api): reorganize result classes into specific packages for better structure
