+++ api/src/main/java/net/onelitefeather/pandorascluster/api/service/result/world/DeleteLandWorldResult.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.pandorascluster.api.service.result;
+package net.onelitefeather.pandorascluster.api.service.result.world;
 
 import net.onelitefeather.pandorascluster.api.land.LandWorld;
